a bunch of changes
also i hate peterson pd. 6 whoever got paint on my hoodie is evil
