+++ app/src/views/MainPage.vue
@@ -1,5 +1,5 @@
 <template>
   <div class="">
-    <h1>ur in</h1>
+    <h1>Main Page</h1>
   </div>
 </template>
